component配下のvcのnavigationItem初期化処理を共通化した
