+++ Pendula/View/Component/001_IrregularLayoutScene/IrregularLayoutViewController.swift
@@ -28,7 +28,7 @@ final class IrregularLayoutViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        setupNavigationItem()
+        setupNavigationItem(title: "001_IrregularLayout")
     }
 
     @IBAction func generate(_ sender: Any) {
@@ -45,22 +45,6 @@ final class IrregularLayoutViewController: UIViewController {
 
     }
 
-}
-
-// TODO: 共通化
-extension IrregularLayoutViewController {
-    private func setupNavigationItem() {
-        navigationItem.title = "001_IrregularLayout"
-        let backButton = UIBarButtonItem(image: R.image.back_arrow()?.withRenderingMode(.alwaysOriginal),
-                                         style: .plain,
-                                         target: self,
-                                         action: #selector(popViewController))
-        navigationItem.leftBarButtonItem = backButton
-    }
-
-    @objc private func popViewController() {
-        navigationController?.popViewController(animated: true)
-    }
 }
 
 // MARK: DataSource
