+++ Pendula/View/Component/002_imos/ImosViewController.swift
@@ -30,7 +30,7 @@ final class ImosViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        setupNavigationItem()
+        setupNavigationItem(title: "002_Imos")
         imosButton.setTitle("Show Result", for: .normal)
     }
 
@@ -59,24 +59,6 @@ final class ImosViewController: UIViewController {
                 cell?.backgroundColor = .darkGray
             }
         }
-    }
-
-}
-
-// TODO: 共通化
-extension ImosViewController {
-
-    private func setupNavigationItem() {
-        navigationItem.title = "002_Imos"
-        let backButton = UIBarButtonItem(image: R.image.back_arrow()?.withRenderingMode(.alwaysOriginal),
-                                         style: .plain,
-                                         target: self,
-                                         action: #selector(popViewController))
-        navigationItem.leftBarButtonItem = backButton
-    }
-
-    @objc private func popViewController() {
-        navigationController?.popViewController(animated: true)
     }
 
 }
